requirements/base.txt
south>=0.8.1
https://github.com/tkaemming/django-orderable/archive/master.tar.gz#egg=django-orderable-1.2.1
django-picklefield>=0.3.0
django-polymodels>=1.2.2